Changed balance to cash in portfolio and add create_cash_position method

+++ db/migrate/20220712161651_create_transactions.rb
@@ -5,9 +5,9 @@ class CreateTransactions < ActiveRecord::Migration[7.0]
       t.date :trade_date
       t.string :symbol
       t.integer :quantity
-      t.decimal :price
-      t.decimal :commission
-      t.decimal :fee
+      t.float :price
+      t.float :commission
+      t.float :fee
 
       t.timestamps
     end
